Movement done!!
the player is moving front and back properly!
also added jump and condition to check isPlayerGrounded
which sets true on OnCollisionEnter() and false on OnCollisionExit
for objects taged as Ground or Box (more may be added)
The player does not move if it is in air

For the player movement
changed playerRb.AddForce() to playerRb.velocity
and now using the playerMesh.transform.forward instead of
camForward.normalized

for jumping, adding a Vector3.up to the current velocity

Major learning!:
in the turret controller previously, bullet parent was set to null
but i commented that in previous commit..
so now now the bullets were spawning in the local space of turret
which made them move in the rotation of turret
so they were kind of following the player, as the turret was rotating
finally uncommented that! and now its working properly!

Signed-off-by: Chinmay Chandak, @CCAtAlvis